requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo[srv]==4.6.1
python-dotenv==1.0.0
certifi==2023.11.17
pyopenssl==23.3.0
gunicorn==21.2.0
reportlab==4.0.7

